+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+#
+- *BREAKING* Rename `BuildOpts` -> `ImageBuildOpts` and `BuildOptsBuilder` -> `ImageBuildOptsBuilder`
+
 # 0.10.0
 - *BREAKING* Use models generated from swagger schema
 - Remove `*Id` and `*IdRef` type aliases and use a common `Id` type for object IDs